cagegorias

+++ src/controllers/offer.controller.ts
@@ -1,3 +1,4 @@
+import { HttpException } from "@/exceptions/httpException";
 import { OfferService } from "@/services/offer.service";
 import {Response, Request, NextFunction} from 'express'
 
@@ -5,6 +6,8 @@ export class OfferController{
     static async getById(req:Request, res:Response, next:NextFunction){
         try{
             const id = Number.parseInt(req.params.id)
+            if (isNaN(id)) throw new HttpException(400, "Invalid offer ID");
+
             // pasar a entero
             const offer = await OfferService.getById(id)
             res.status(200).json(offer)
@@ -15,7 +18,8 @@ export class OfferController{
 
     static async getAll(req:Request, res:Response, next: NextFunction){
         try{
-            const user = await OfferService.getAll()
+            const { title } = req.query;
+            const user = await OfferService.getAll(title as string)
             res.status(200).json(user)
         }catch(error){
             next(error)
@@ -27,6 +31,9 @@ export class OfferController{
             const offerData = req.body
             const userId = req.body.user.id
 
+            if (!userId) throw new HttpException(400, "User creator ID is required");
+            
+
             const newOffer = await OfferService.create(userId, offerData)
             res.status(200).json(newOffer)
         }catch(error){
@@ -37,6 +44,8 @@ export class OfferController{
         try{
             const offerData = req.body
             const id = Number.parseInt(req.params.id)
+            if (isNaN(id)) throw new HttpException(400, "Invalid offer ID");
+
             const updatedOffer = await OfferService.update(id, offerData)
             res.status(200).json(updatedOffer)
         }catch(error){
@@ -47,6 +56,8 @@ export class OfferController{
     static async delete(req:Request, res:Response, next: NextFunction){
         try{
             const id = Number.parseInt(req.params.id)
+            if (isNaN(id)) throw new HttpException(400, "Invalid offer ID");
+
             const deletedOffer = await OfferService.delete(id)
             res.status(200).json(deletedOffer)
         }catch(error){
@@ -56,6 +67,8 @@ export class OfferController{
     static async rate(req:Request, res:Response, next: NextFunction){
         try{
             const id = Number.parseInt(req.params.id)
+            if (isNaN(id)) throw new HttpException(400, "Invalid offer ID");
+
             const {value} = req.body
             const userId = req.body.user.id
 
@@ -69,6 +82,8 @@ export class OfferController{
     static async getRate(req:Request, res:Response, next: NextFunction){
         try{
             const id = Number.parseInt(req.params.id)
+            if (isNaN(id)) throw new HttpException(400, "Invalid offer ID");
+
             await OfferService.getRate(id)
             res.status(200).json({message: 'Offer rate successfully'})
         }catch(error){
